Add Support for reading Id from delta resources with key properties set.
